Create report.js
This is the original JAVASCRIPT with an IE8 COMPATIBILITY ISSUE for the "DOWNLOAD .CSV FILE BUTTON" The warning alert shows line #56 "CONSOLE" call conflicting. . . "CONSOLE" is for DEBUGGING PURPOSES and doesn't exist in IE8-- should be removed for PRODUCTION. GREAT TIP - Leave the SCRIPT ALERTS ON in the IE ADVANCED SETTINGS so while testing an alert will pop up.